feat(feature): decouple django registration from admin definitions
