--- events/admin/__init__.py
@@ -1,7 +1,0 @@
-from events.admin.car import CarAdmin
-from events.admin.competitor import CompetitorAdmin
-from events.admin.event import EventAdmin
-from events.admin.inscription import InscriptionAdmin
-from events.admin.team import TeamAdmin
-
-__all__ = ["EventAdmin", "CompetitorAdmin", "CarAdmin", "TeamAdmin", "InscriptionAdmin"]
