+++ events/admins/team.py
@@ -1,0 +1,5 @@
+from unfold.admin import ModelAdmin
+
+
+class TeamAdmin(ModelAdmin):
+    list_display = ["name"]
